refactor: Customer 멤버필드 타입 변경
